Tidy up duplicate stock definitions
